+++ assets/controllers/widget_controller.js
@@ -1,17 +1,44 @@
-import {Controller} from '@hotwired/stimulus';
+import {Controller} from "@hotwired/stimulus";
 
 export default class extends Controller {
-	static targets = ["modal"];
+    static targets = ["modal"];
 
-	connect() {
-		super.connect();
-	}
+    connect() {
+        super.connect();
+        this.toggleInitialVisibility();
+    }
 
-	open() {
-		this.modalTarget.classList.remove('hidden');
-	}
+    toggle(event) {
+        // Check the state of the checkbox (useLDAP - Admin Page SamlProvider add/edit)
+        const isChecked = event.currentTarget.checked;
 
-	close() {
-		this.modalTarget.classList.add('hidden');
-	}
+        if (isChecked) {
+            this.open();
+        } else {
+            this.close();
+        }
+    }
+
+    toggleInitialVisibility() {
+        // Ensure the state is correct on page load
+        const checkbox = this.element.querySelector('input[type="checkbox"]');
+        if (!checkbox) {
+            return;
+        }
+
+        console.log("Toggle Initial Visibility - Checkbox checked state:", checkbox.checked);
+        if (checkbox.checked) {
+            this.open();
+        } else {
+            this.close();
+        }
+    }
+
+    open() {
+        this.modalTarget.classList.remove("hidden");
+    }
+
+    close() {
+        this.modalTarget.classList.add("hidden");
+    }
 }