+++ src/Command/LDAPSyncCommand.php
@@ -2,9 +2,10 @@
 
 namespace App\Command;
 
+use App\Entity\SamlProvider;
 use App\Enum\UserProvider;
 use App\Enum\UserRadiusProfileRevokeReason;
-use App\Repository\SettingRepository;
+use App\Repository\SamlProviderRepository;
 use App\Repository\UserExternalAuthRepository;
 use App\Repository\UserRepository;
 use App\Service\ProfileManager;
@@ -22,9 +23,9 @@ class LDAPSyncCommand extends Command
 {
     public function __construct(
         private readonly UserRepository $userRepository,
-        private readonly SettingRepository $settingRepository,
         private readonly ProfileManager $profileManager,
-        private readonly UserExternalAuthRepository $userExternalAuthRepository
+        private readonly UserExternalAuthRepository $userExternalAuthRepository,
+        private readonly SamlProviderRepository $samlProviderRepository,
     ) {
         parent::__construct();
     }
@@ -36,10 +37,36 @@ class LDAPSyncCommand extends Command
     protected function execute(InputInterface $input, OutputInterface $output): int
     {
         $io = new SymfonyStyle($input, $output);
-        if ($this->settingRepository->findOneBy(['name' => 'SYNC_LDAP_ENABLED'])->getValue() === 'false') {
-            $io->writeln('LDAP sync is disabled');
-            return Command::SUCCESS;
+        $samlProvider = $this->samlProviderRepository->findOneBy([
+            'isActive' => true,
+            'isLDAPActive' => true,
+            'deletedAt' => null
+        ]);
+
+        if (!$samlProvider) {
+            $io->writeln(
+                '<error>No active SAML Provider is currently configured. ' .
+                ' Please ensure a SAML Provider is set up and associated with an active LDAP Credential.</error>'
+            );
+            return Command::FAILURE;
         }
+
+        // Check if the SAML Provider has valid LDAP configuration
+        if (
+            !$samlProvider->getLdapServer() ||
+            !$samlProvider->getLdapBindUserDn() ||
+            !$samlProvider->getLdapBindUserPassword()
+        ) {
+            $io->writeln(
+                sprintf(
+                    '<error>No valid LDAP configuration is associated with the active SAML Provider (%s).</error>',
+                    $samlProvider->getName()
+                )
+            );
+            return Command::FAILURE;
+        }
+
+
         $ldapEnabledUsers = $this->userRepository->findLDAPEnabledUsers();
         $io->writeln('Found ' . count($ldapEnabledUsers) . ' LDAP enabled users');
         foreach ($ldapEnabledUsers as $user) {
@@ -50,7 +77,7 @@ class LDAPSyncCommand extends Command
                     $providerId = $externalAuth->getProviderId();
                     $io->writeln('Syncing ' . $providerId . ' with LDAP');
 
-                    $ldapUser = $this->fetchUserFromLDAP($providerId);
+                    $ldapUser = $this->fetchUserFromLDAP($providerId, $samlProvider);
                     if (is_null($ldapUser)) {
                         $io->writeln('User ' . $providerId . ' not found in LDAP, disabling');
                         $this->profileManager->disableProfiles(
@@ -87,11 +114,11 @@ class LDAPSyncCommand extends Command
         return Command::SUCCESS;
     }
 
-    private function fetchUserFromLDAP(string $identifier)
+    private function fetchUserFromLDAP(string $identifier, SamlProvider $samlProvider)
     {
-        $ldapServer = $this->settingRepository->findOneBy(['name' => 'SYNC_LDAP_SERVER'])->getValue();
-        $ldapUsername = $this->settingRepository->findOneBy(['name' => 'SYNC_LDAP_BIND_USER_DN'])->getValue();
-        $ldapPassword = $this->settingRepository->findOneBy(['name' => 'SYNC_LDAP_BIND_USER_PASSWORD'])->getValue();
+        $ldapServer = $samlProvider->getLdapServer();
+        $ldapUsername = $samlProvider->getLdapBindUserDn();
+        $ldapPassword = $samlProvider->getLdapBindUserPassword();
         $ldapConnection = ldap_connect($ldapServer) or die("Could not connect to LDAP server.");
         ldap_set_option($ldapConnection, LDAP_OPT_DEREF, LDAP_DEREF_ALWAYS);
         ldap_set_option($ldapConnection, LDAP_OPT_PROTOCOL_VERSION, 3);
@@ -102,9 +129,9 @@ class LDAPSyncCommand extends Command
         $searchFilter = str_replace(
             "@ID",
             $identifier,
-            $this->settingRepository->findOneBy(['name' => 'SYNC_LDAP_SEARCH_FILTER'])->getValue()
+            $samlProvider->getLdapSearchFilter()
         );
-        $searchBaseDN = $this->settingRepository->findOneBy(['name' => 'SYNC_LDAP_SEARCH_BASE_DN'])->getValue();
+        $searchBaseDN = $samlProvider->getLdapSearchBaseDn();
         $searchResult = ldap_search(
             $ldapConnection,
             $searchBaseDN,