+++ src/Form/SamlProviderType.php
@@ -9,12 +9,16 @@ use App\Validator\SAMLProviderUrl;
 use App\Validator\X509Certificate;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\Form\AbstractType;
+use Symfony\Component\Form\Extension\Core\Type\CheckboxType;
+use Symfony\Component\Form\Extension\Core\Type\PasswordType;
 use Symfony\Component\Form\Extension\Core\Type\TextareaType;
 use Symfony\Component\Form\Extension\Core\Type\TextType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 use Symfony\Component\Validator\Constraints as Assert;
+use Symfony\Component\Validator\Constraints\Callback;
 use Symfony\Component\Validator\Constraints\NotBlank;
+use Symfony\Component\Validator\Context\ExecutionContextInterface;
 
 class SamlProviderType extends AbstractType
 {
@@ -31,11 +35,11 @@ class SamlProviderType extends AbstractType
                 'empty_data' => '',
                 'constraints' => [
                     new NotBlank([
-                        'message' => 'Please enter a valid provider name'
+                        'message' => 'Please enter a valid provider name.'
                     ]),
                     new CamelCase(),
+                    // Ensure 'name' is unique
                     new Assert\Callback(function ($value, $context): void {
-                        // Check for uniqueness of the 'name' field in the database
                         $existingProvider = $this->entityManager->getRepository(SamlProvider::class)
                             ->findOneBy(['name' => $value]);
 
@@ -53,10 +57,10 @@ class SamlProviderType extends AbstractType
                 'empty_data' => '',
                 'constraints' => [
                     new NotBlank([
-                        'message' => 'Please enter a valid provider IDP Entity ID'
+                        'message' => 'Please enter a valid SAML IDP Entity ID.'
                     ]),
+                    // Ensure 'idpEntityId' is unique
                     new Assert\Callback(function ($value, $context): void {
-                        // Validation for uniqueness of 'idpEntityId'
                         $existingEntity = $this->entityManager->getRepository(SamlProvider::class)
                             ->findOneBy(['idpEntityId' => $value]);
 
@@ -75,10 +79,10 @@ class SamlProviderType extends AbstractType
                 'empty_data' => '',
                 'constraints' => [
                     new NotBlank([
-                        'message' => 'Please enter a valid provider IDP SSO URL'
+                        'message' => 'Please enter a valid SAML IDP SSO URL.'
                     ]),
                     new Assert\Callback(function ($value, $context): void {
-                        // Validation for uniqueness of 'idpEntityId'
+                        // Ensure 'idpSsoUrl' is unique
                         $existingEntity = $this->entityManager->getRepository(SamlProvider::class)
                             ->findOneBy(['idpSsoUrl' => $value]);
 
@@ -90,7 +94,7 @@ class SamlProviderType extends AbstractType
                         }
                     }),
                     new Assert\Url([
-                        'message' => 'The value {{ value }} is not a valid URL.',
+                        'message' => 'The value "{{ value }}" is not a valid URL.',
                         'protocols' => ['http', 'https'],
                     ]),
                 ],
@@ -100,10 +104,10 @@ class SamlProviderType extends AbstractType
                 'empty_data' => '',
                 'constraints' => [
                     new NotBlank([
-                        'message' => 'Please enter a valid provider SP Entity ID'
+                        'message' => 'Please enter a valid SAML SP Entity ID.'
                     ]),
                     new Assert\Callback(function ($value, $context): void {
-                        // Validation for uniqueness of 'idpEntityId'
+                        // Ensure 'spEntityId' is unique
                         $existingEntity = $this->entityManager->getRepository(SamlProvider::class)
                             ->findOneBy(['spEntityId' => $value]);
 
@@ -122,14 +126,14 @@ class SamlProviderType extends AbstractType
                 'empty_data' => '',
                 'constraints' => [
                     new NotBlank([
-                        'message' => 'Please enter a valid provider SP ACS URL'
+                        'message' => 'Please enter a valid SAML SP ACS URL.'
                     ]),
                     new Assert\Url([
-                        'message' => 'The value {{ value }} is not a valid URL.',
+                        'message' => 'The value "{{ value }}" is not a valid URL.',
                         'protocols' => ['http', 'https'],
                     ]),
                     new Assert\Callback(function ($value, $context): void {
-                        // Validation for uniqueness of 'idpEntityId'
+                        // Ensure 'spAcsUrl' is unique
                         $existingEntity = $this->entityManager->getRepository(SamlProvider::class)
                             ->findOneBy(['spAcsUrl' => $value]);
 
@@ -147,9 +151,69 @@ class SamlProviderType extends AbstractType
                 'empty_data' => '',
                 'constraints' => [
                     new NotBlank([
-                        'message' => 'Please enter a valid provider IDP X509 certificate'
+                        'message' => 'Please enter a valid SAML IDP X509 certificate.'
                     ]),
                     new X509Certificate()
+                ],
+            ])
+            ->add('isLDAPActive', CheckboxType::class, [
+                'label' => 'Enable LDAP',
+                'mapped' => true,
+                'required' => false,
+            ])
+            ->add('ldapServer', TextType::class, [
+                'label' => 'LDAP Server',
+                'required' => false,
+                'constraints' => [
+                    new Callback(function ($value, ExecutionContextInterface $context): void {
+                        $formData = $context->getRoot()->getData();
+                        if (empty($value) && $formData->getIsLDAPActive()) {
+                            $context->buildViolation('Please enter a valid LDAP Server.')
+                                ->atPath('ldapServer')
+                                ->addViolation();
+                        }
+                    }),
+                ],
+                'attr' => [
+                    'autocomplete' => 'off',
+                ],
+            ])
+            ->add('ldapBindUserDn', TextType::class, [
+                'label' => 'Bind User DN',
+                'required' => false,
+                'constraints' => [
+                    new Callback(function ($value, ExecutionContextInterface $context): void {
+                        $formData = $context->getRoot()->getData();
+                        if (empty($value) && $formData->getIsLDAPActive()) {
+                            $context->buildViolation('Please enter a valid Bind Distinguished Name.')
+                                ->atPath('ldapBindUserDn')
+                                ->addViolation();
+                        }
+                    }),
+                ],
+                'attr' => [
+                    'autocomplete' => 'off',
+                ],
+            ])
+            ->add('ldapBindUserPassword', PasswordType::class, [
+                'label' => 'Bind User Password',
+                'required' => false,
+                'attr' => [
+                    'autocomplete' => 'off',
+                ],
+            ])
+            ->add('ldapSearchBaseDn', TextType::class, [
+                'label' => 'Search Base DN',
+                'required' => false,
+                'attr' => [
+                    'autocomplete' => 'off',
+                ],
+            ])
+            ->add('ldapSearchFilter', TextType::class, [
+                'label' => 'Search Filter',
+                'required' => false,
+                'attr' => [
+                    'autocomplete' => 'off',
                 ],
             ]);
     }